backend/requirements.txt
# EAB Enzyme Discovery System - Python Dependencies

# Bioinformatics
biopython>=1.81

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
# sqlite3 is included in Python standard library

# HTTP requests
requests>=2.31.0

# Utilities
tqdm>=4.65.0

# Optional: Web interface
streamlit>=1.28.0

# Optional: Excel export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
